requirements.txt
selenium==4.1.0
pytest==6.2.5
python-dotenv==0.20